refactor record trace step helpers
